refactor middleware folder
