+++ server/requests/get_single_user.rest
@@ -1,0 +1,1 @@
+GET http://localhost:3001/api/users/tony_stark
