+++ server/utils/config.js
@@ -6,4 +6,4 @@ const DB_NAME = process.env.DB_NAME;
 module.exports = {
     DB_SERVER,
     DB_NAME
-}
+};
